Don't do reverse DNS lookups on IP addresses
Reverse DNS lookups always fail on certain WAN IPs, which prevents game rooms associated to them from being broadcast in Warhawk's LAN lobby.

This resolves that issue by replacing Dns.GetHostEntry() with Dns.GetHostAddresses(). The latter only performs a reverse DNS lookup when a host name is passed to it. If an IP address is provided, it returns the IP as-is.

Fixes #4.