Ping PS3 before sending game room details
When a user enters or refreshes Warhawk's LAN lobby, this pings their PS3 in order to create an ARP entry for it. All room details are subsequently sent out.

This resolves a common issue in which only one room appears in the lobby (unless the user manually refreshes). This adjustment will cause all rooms to always show up.

Fixes #1.